chore: rename session persistence policy to session affinity
