+++ crates/agentgateway/Cargo.toml
@@ -149,6 +149,7 @@ tower-serve-static.workspace = true
 tower.workspace = true
 tracing-subscriber.workspace = true
 tracing.workspace = true
+twox-hash.workspace = true
 typespec_client_core.workspace = true
 url.workspace = true
 uuid.workspace = true
